# docs/tenant-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "operationId": "ReviewsController_getReviews",
        "summary": "Get reviews",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Daily",
                "Weekly"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews filtered by type"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "ReviewsController_createReview",
        "summary": "Create review",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/reviews/summary": {
      "get": {
        "operationId": "ReviewsController_getSummary",
        "summary": "Get review summary for dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Review dashboard summary"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/sales/planning": {
      "get": {
        "operationId": "SalesController_getSalesPlanning",
        "summary": "Get sales planning for year",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "example": 2025,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales planning configuration"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "SalesController_upsertSalesPlanning",
        "summary": "Create or update sales planning",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSalesPlanningDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sales planning saved"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/sales/tracker": {
      "get": {
        "operationId": "SalesController_getSalesTracker",
        "summary": "Get sales tracker for month",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2025-10",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly sales tracker"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "SalesController_upsertSalesTracker",
        "summary": "Create or update sales tracker",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSalesTrackerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sales tracker saved"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/sales/summary": {
      "get": {
        "operationId": "SalesController_getSummary",
        "summary": "Get sales dashboard summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sales dashboard summary"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/sales/trackers": {
      "get": {
        "operationId": "SalesController_getAllSalesTrackers",
        "summary": "Get all sales trackers with pagination",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query string",
            "schema": {
              "example": "revenue",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field and direction (e.g., createdAt:desc, name:asc)",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of sales trackers"
          },
          "400": {
            "description": "Bad Request – validation failed for payload or query parameters"
          },
          "401": {
            "description": "Unauthorized – missing or invalid JWT"
          },
          "403": {
            "description": "Forbidden – insufficient role or tenant scope"
          }
        },
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "BG Accountability Tenant API",
    "description": "REST API for the BG Accountability Platform - Tenant Services.\n\n## Overview\nThis API serves both the Tenant Web Dashboard (Next.js) and Mobile App (Flutter).\n\n## Authentication\nAll protected endpoints require a valid JWT Bearer token.\n- Login via `POST /api/v1/auth/login`\n- Include token in header: `Authorization: Bearer <token>`\n- Refresh tokens via `POST /api/v1/auth/refresh`\n\n## Multi-Factor Authentication (MFA)\n- If MFA is enabled, login returns a `tempToken`\n- Complete MFA via `POST /api/v1/auth/mfa/login` with tempToken + TOTP code\n\n## Rate Limiting\n- Default: 100 requests per minute per IP\n- Auth endpoints: 10 requests per minute per IP\n\n## Error Format\nAll errors follow: `{ success: false, message: string, code?: string, details?: object }`\n\n## Pagination\nPaginated endpoints use: `{ items: T[], total: number, page: number, pageSize: number }`\n",
    "version": "1.0.0",
    "contact": {
      "name": "BG Accountability",
      "url": "https://bridgegaps.app",
      "email": "support@bridgegaps.app"
    },
    "license": {
      "name": "Proprietary",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local Development"
    },
    {
      "url": "https://api.bridgegaps.app",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "Daily",
              "Weekly"
            ]
          },
          "mood": {
            "type": "number"
          },
          "content": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
      "UpsertSalesPlanningDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number"
          },
          "q1": {
            "type": "number"
          },
          "q2": {
            "type": "number"
          },
          "q3": {
            "type": "number"
          },
          "q4": {
            "type": "number"
          }
        },
        "required": [
          "year"
        ]
      },
      "UpsertSalesTrackerDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string"
          },
          "target": {
            "type": "number"
          },
          "achieved": {
            "type": "number"
          },
          "mtd": {
            "type": "number"
          },
          "ytd": {
            "type": "number"
          },
          "orders": {
            "type": "number"
          },
          "asp": {
            "type": "number"
          },
          "expenses": {
            "type": "number"
          },
          "profit": {
            "type": "number"
          }
        },
        "required": [
          "month"
        ]
      },
...
    }
  }
}